CHROMIUM: char: tpm: need to resume powered-off TPMs after resume
I was looking at getting rid of some of our tpm patches, and this was
the remaining delta. I asked Luigi about it, and he said (in the below
referenced bug):

"The code that this patch removes is needed by devices that power off
the TPM during suspend.  I believe Alex and ZGB work that way, i.e. they
need the code that this patch removes.

If the platform maintains power to the TPM during suspend, then this
code isn't needed, i.e. you may apply the diff.  But since we're still
supporting Alex and ZGB, that code needs to be there."

So, adding it back without keeping the full add-removesome-addsome
history of the TPM drivers.

BUG=chromium-os:27576
TEST=regression suite

Change-Id: Idae59dcdf9fc66b14a522715824c9d227fdd375e
[sonnyrao: 3.8: fixed conflicts]
[sonnyrao: 3.8: squashed deadlock patch]
[sonnyrao: 3.8: squashed reconcile mainline patch]
[sonnyrao: 3.8: squashed cleanup missing newlines]
[sonnyrao: 3.8: squashed always run ContinueSelftest]
Signed-off-by: Olof Johansson <olofj@chromium.org>