UPSTREAM: usb-serial: ftdi_sio: fix oops during autosuspend
commit 5cbe61c5aff0a8ada691eb8b07dbfb55c303f640 upstream.

This patch (as1550) fixes a bug in the usb-serial core that affects
the ftdi_sio driver and most likely others as well.  The core
implements suspend and resume routines, but it doesn't store pointers
to those routines in the usb_driver structures that it registers,
even though it does set those drivers' supports_autosuspend flag.  The
end result is that when one of these devices is autosuspended, we try
to call through a NULL pointer.

The patch fixes the problem by setting the suspend and resume method
pointers to the appropriate routines in the USB serial core, along
with the supports_autosuspend field, in each driver as it is
registered.

This should be back-ported to all the stable kernels that have the new
usb_serial_register_drivers() interface.

Signed-off-by: Alan Stern <stern@rowland.harvard.edu>
Reported-and-tested-by: Frank Schäfer <schaefer.frank@gmx.net>
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>

(cherry picked from commit 2aa5c3521c154db6e3bbbda1a2f0b07da4938626)
Cherry picked from upstream's stable/linux-stable.git/linux-3.4.y

BUG=chromium:263565
TEST=Plug a Keyspan USB-to-serial adapter into a stumpy, watch it stay
alive for 5+ seconds.

Change-Id: Ib04984e6664870aa278f4edf472eb130cc9fc254
Signed-off-by: Julius Werner <jwerner@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/63286
Reviewed-by: Vincent Palatin <vpalatin@chromium.org>