CHROMIUM: ARM: enable xclkout for audio on daisy/snow
Daisy/snow need xclkout to be enabled since it's their reference clock
for the audio codec. Enabling the output on smdk shouldn't be
harmful so just do it globally for now.

Note that this should be moved over to be a regular system clock instead
of just being forced on like this.

Change-Id: Ia4f612ff56c6715aa99e9b50e51b5e03874feb8a
Signed-off-by: Olof Johansson <olofj@chromium.org>
Reviewed-on: https://gerrit-int.chromium.org/20545
Reviewed-by: Dylan Reid <dgreid@chromium.org>
Commit-Ready: Olof Johansson <olofj@google.com>
Tested-by: Olof Johansson <olofj@google.com>