HACK: ARM: exynos: WiFi power-on sequence
Manually sequence the PDn and RESETn lines to match the mlw8797 spec.
The sequence now waits for SLP_CLK to settle before raising RESETn,
waiting 1.5ms (min time specified is 1ms), and then raising PDn.
Remove the explicit BT reset as the BT and WiFi reset lines are tied.

BUG=None
TEST=Boot; WiFi card is now detected

[dianders: pulled this up to 3.8 and merged with a few future changes;
at some point in time we need to find a better place for this]

Change-Id: I8128c8a8d32db1b449dd1e26f5b2698c45fe7cd8
Signed-off-by: Sam Leffler <sleffler@chromium.org>
Signed-off-by: Doug Anderson <dianders@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/47479
Reviewed-by: Olof Johansson <olofj@chromium.org>