CHROMIUM: drm/exynos: return an error when flip called with event flag
The event flag to the page_flip ioctl is optional. If it is set,
an event will be sent to the drm fd after page flip is complete.
Otherwise, no event is sent after page flip. The exynos implementation
was doing nothing if the event flag was not set. This would
break any application expecting a page flip to happen. Rather than
do nothing (which is incorrect behavior), we'll now return and log
an error so that we can detect anything that's broken.

This also removes one level of indenting, making the code easier to
read.

BUG=chrome-os-partner:18783
TEST=Verified that no errors were logged.

Signed-off-by: Mandeep Singh Baines <msb@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/28853

Change-Id: I2e7c4d55c559ce4d9f92fe2089e23e3e30ce2441
Signed-off-by: Stéphane Marchesin <marcheu@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/48336
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Daniel Kurtz <djkurtz@chromium.org>