CHROMIUM: gpio: samsung: add flags for suspend conditions for gpios
On suspend, gpio lines are driven active low by default.  This
extends the flags field for samsung,exynos4-gpio compatible
gpio chips to support specifying the suspend state of its gpios.

In addition modifying the saving of GPIO registers to only save
the CONPDN and PUDPDN registers for chips that indicate they support
it.

The information is passed through a new powerdown-settings property
in the GPIO chip configuration.  It specifies the state during powerdown
of each individual gpio as:
     0 - Use existing
     1 - Float (input, no PU/PD)
     2 - Pull up (input)
     3 - Pull down (input)
     4 - Output high
     5 - Output low

BUG=chrome-os-partner:11846
TEST=Suspend/resume multiple times, confirm state of pins
  from interaction with other components.
  Insert bad values, confirm proper error handling.
Signed-off-by: Jonathan Kliegman <kliegs@chromium.org>

Change-Id: Ife5e9070eaaf1fb5feec2d70623fddc5aea48613
Reviewed-on: https://gerrit.chromium.org/gerrit/28639
Reviewed-by: Olof Johansson <olofj@chromium.org>
Commit-Ready: Jon Kliegman <kliegs@chromium.org>
Tested-by: Jon Kliegman <kliegs@chromium.org>