CHROMIUM: gpu: vithar: tune GPU clock governor
Address some issues in the GPU frequency selection algorithm.

The main change is to use a weighted average of the last 4 samples of the
GPU utilisation value to drive frequency step up/down's and do it based
on the number of samples instead of consecutive out-of-range samples.
This smooths transitions where the previous algorithm would not transition
because of a single sample that was out-of-range (which can happen for
various reasons outside the GPU).  Four samples is chosen as a compromise
between smoothing bumps and quickly identifying changes.

Increase the sampling frequency and tune the up/down step thresholds to
be more aggresive up-shifting and to happen faster.  The new values are
optimized for performance (over power) based on a set of workloads.

Remove the mechanism that caused a clock boost when the GPU goes !busy
after a period of being idle.  This caused anomalies such as clock ramping
at the login screen due to periodic short updates (e.g. to re-draw the
clock face).  Instead of we depend on the normal clock ramp up mechanism
which has been tuned for faster response.

Remove the special handling to speed up-shifting when seeing 100%
utilisation samples.  This happens rarely (based on stats collected) and
is now handled by more frequent sampling and shorter up-shift thresholds.

While here: move the up/down thresholds to the infotbl (simplifies
code), change the up/down thresholds to be the minimum value required
(> replaced by >=), simplify utilisation calculation, and fix a bunch
of checkpatch complaints.

Signed-off-by: Sam Leffler <sleffler@chromium.org>

BUG=chrome-os-partner:13355
TEST=open multiple webgl demo's, video playback, static html content, and a new tab; monitor gpu clock to verify the algorithm chooses an "optimal clock" (relative to performance w/o dvfs); switch between tabs and verify the clock is changed "quickly" and there is no instability in the frequency selection

Change-Id: I826e574d9215147399eb4c49844fbcd9500c3273
Reviewed-on: https://gerrit.chromium.org/gerrit/31745
Reviewed-by: Bryan Freed <bfreed@chromium.org>
Tested-by: Sam Leffler <sleffler@chromium.org>
Commit-Ready: Sam Leffler <sleffler@chromium.org>