CHROMIUM: DMA: PL330: Report residue for dev to mem transfers.
When setting the residue, account for device to memory transfers.  Do
this by checking both the source and destination addresses against the
active descriptors.  For memory to device the source address will be
incrementing, for device to memory the destination will increment and
for memory to memory both will increment.

BUG=chrome-os-partner:12776
TEST=enter hangout, ask if other end can hear you clearly.
'arecord -B10000 -f data /tmp/asdf.wav' then playback.

Change-Id: I4debd2e8b6ea1c2c9842e692016e83d7c852de91
Signed-off-by: Dylan Reid <dgreid@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/33922
Reviewed-by: Olof Johansson <olofj@chromium.org>