samsung: pm-check: Add interleaving support
When pm-check is used in conjunction with bitfix we can get much
faster processing times if we change the order that we process chunks.
Add support for changing the order that we process pages in save so
that bitfix can use it.

BUG=chrome-os-partner:15655
TEST=Timed before and after and saw 500-600ms speedup.  See associated
bitfix change (dd13e37d3ca428eb89ee6d9afe3b258ec4ca0bfd) for details.

Change-Id: I46156f03d0689ba1cee3398bf16de34b06b6bce0
Original-Change-Id: If2b720a75f5274df0e5dc11d77662d73a4581ffb
Signed-off-by: Doug Anderson <dianders@chromium.org>
Originally-Reviewed-on: https://gerrit.chromium.org/gerrit/37690
Reviewed-by: Jon Kliegman <kliegs@chromium.org>
Signed-off-by: Olof Johansson <olofj@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/171955