UPSTREAM: arm: arch_timer: split cntfrq accessor
The CNTFRQ register is not duplicated for physical and virtual timers,
and accessing it as if it were is confusing.

Instead, use a separate accessor which doesn't take the access type
as a parameter.

BUG=chrome-os-partner:13805
TEST=Built and booted the kernel on pi

Signed-off-by: Mark Rutland <mark.rutland@arm.com>
Acked-by: Catalin Marinas <catalin.marinas@arm.com>
Acked-by: Marc Zyngier <marc.zyngier@arm.com>
Acked-by: Santosh Shilimkar <santosh.shilimkar@ti.com>
Signed-off-by: Chirantan Ekbote <chirantan@chromium.org>
(cherry picked from commit fd5583a4c271ec03e2da04196aaaab177b385eb8)

Change-Id: I300aa48768306106c358d4714a5d1a1e8b23932f
Reviewed-on: https://chromium-review.googlesource.com/191584
Reviewed-by: Olof Johansson <olofj@chromium.org>
Tested-by: Chirantan Ekbote <chirantan@chromium.org>
Commit-Queue: Chirantan Ekbote <chirantan@chromium.org>