UPSTREAM: arm: arch_timer: divorce from local_timer api
Currently, the arch_timer driver is tied to the arm port, as it relies
on code in arch/arm/smp.c to setup and teardown timers as cores are
hotplugged on and off. The timer is registered through an arm-specific
registration mechanism, preventing sharing the driver with the arm64
port.

This patch moves the driver to using a cpu notifier instead, making it
easier to port.

BUG=chrome-os-partner:13805
TEST=Built and booted the kernel on pi

Signed-off-by: Mark Rutland <mark.rutland@arm.com>
Acked-by: Catalin Marinas <catalin.marinas@arm.com>
Acked-by: Marc Zyngier <marc.zyngier@arm.com>
Acked-by: Santosh Shilimkar <santosh.shilimkar@ti.com>
Signed-off-by: Chirantan Ekbote <chirantan@chromium.org>
(cherry picked from commit 1ba1cefc277865a0ac222f53bbbf2ebacad1559a)

Change-Id: I83eac192420f555d3a8f2ce06142910eef573eca
Reviewed-on: https://chromium-review.googlesource.com/191587
Reviewed-by: Olof Johansson <olofj@chromium.org>
Tested-by: Chirantan Ekbote <chirantan@chromium.org>
Commit-Queue: Chirantan Ekbote <chirantan@chromium.org>