UPSTREAM: arm: arch_timer: add arch_counter_set_user_access
Several bits in CNTKCTL reset to 0, including PL0VTEN. For architectures
using the generic timer which wish to have a fast gettimeofday vDSO
implementation, these bits must be set to 1 by the kernel. For
architectures without a vDSO, it's best to leave the bits set to 0 for
now to ensure that if and when support is added, it's implemented sanely
architecture wide.

As the bootloader might set PL0VTEN to a value that doesn't correspond
to that which the kernel prefers, we must explicitly set it to the
architecture port's preferred value.

This patch adds arch_counter_set_user_access, which sets the PL0 access
permissions to that required by the architecture. For arch/arm, this
currently means disabling all userspace access.

BUG=chrome-os-partner:13805
TEST=Built and booted the kernel on pi

Signed-off-by: Mark Rutland <mark.rutland@arm.com>
Acked-by: Catalin Marinas <catalin.marinas@arm.com>
Signed-off-by: Chirantan Ekbote <chirantan@chromium.org>
(cherry picked from commit b2deabe3ba664a1ec47400c0ca285e951874e0cc)

Change-Id: Ia67612a987c63b409e69b6dc61f2ea989ab5379b
Reviewed-on: https://chromium-review.googlesource.com/191588
Reviewed-by: Olof Johansson <olofj@chromium.org>
Tested-by: Chirantan Ekbote <chirantan@chromium.org>
Commit-Queue: Chirantan Ekbote <chirantan@chromium.org>