CHROMIUM: gobi: Remove leaky replacement of fops in devqmi_write
The devqmi_write function sometimes overwrites devqmi_fops with
def_chr_fops. When that happens, the devqmi_fops reference is leaked,
and the file is never cleaned up because devqmi_release is not called.

It's not clear what this is for, and it looks like the other file
callbacks already check device_valid() before doing anything.
Therefore we can probably just remove this.

BUG=chrome-os-partner:16158
TEST=suspend_stress_test; stop cromo; cat /sys/module/gobi/refcnt

Change-Id: Ie8aecf98784165629864c90b60ee1c20d2106bef
Signed-off-by: Michael Spang <spang@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/38058
Reviewed-by: Vincent Palatin <vpalatin@chromium.org>
Reviewed-by: Mandeep Singh Baines <msb@chromium.org>