drm/exynos: Remove the i2c drivers and use devtree
The i2c client was previously being passed into the hdmi driver via a
dedicated i2c driver, and then a global variable. This patch removes all
of that and just uses the device tree to get the i2c_client. This patch
also properly references the client so we don't lose it before we're
done with it.

BUG=None
TEST=Tested boot & hotplug, no regressions found

Change-Id: I11b8565cecf6c8290851339f923c46ae5513dc74
Signed-off-by: Sean Paul <seanpaul@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/38940
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>