BACKPORT: rcu: Silence compiler array out-of-bounds false positive
It turns out that gcc 4.8 warns on array indexes being out of bounds
unless it can prove otherwise.  It gives this warning on some RCU
initialization code.  Because this is far from any fastpath, add
an explicit check for array bounds and panic if so.  This gives the
compiler enough information to figure out that the array index is never
out of bounds.

However, if a similar false positive occurs on a fastpath, it will
probably be necessary to tell the compiler to keep its array-index
anxieties to itself.  ;-)

Markus Trippelsdorf <markus@trippelsdorf.de>
Signed-off-by: Paul E. McKenney <paulmck@linux.vnet.ibm.com>
Reviewed-by: Josh Triplett <josh@joshtriplett.org>

(cherry picked from commit 4930521ae10fd28ebc713107fd243c8044a95415)
Signed-off-by: Yunlian Jiang <yunlian@google.com>

BUG=chromium:392918
TEST=emerge-x86-alex chromeos-kernel-3_8 with gcc 4.9 passes.

Change-Id: Iefe6e64b37db5e0f5dd6235ed7af4fad3ea1071c
Reviewed-on: https://chromium-review.googlesource.com/207380
Reviewed-by: Olof Johansson <olofj@chromium.org>
Commit-Queue: Yunlian Jiang <yunlian@chromium.org>
Tested-by: Yunlian Jiang <yunlian@chromium.org>