KDS: fix BUG in change I4df35f08
In the implementation of change I4df35f08 the same name
is used for the macro "pos" parameter and "head" parameter.
As consequence "head" gets corrupted. this fix introduces an
auxiliar pointer for "head" and rearrange the code to make it
a bit more readable

BUG=chromium-os:36071,chrome-os-partner:15349,chrome-os-partner:14965
TEST=Tested on Daisy Board
Signed-off-by: Gabriele Paoloni <gabriele.paoloni@arm.com>

Change-Id: Icee1c07c9a87f137cdfd87fb1fe04a83f41a44b1
Reviewed-on: https://gerrit.chromium.org/gerrit/39089
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>
Reviewed-by: Mandeep Singh Baines <msb@chromium.org>
Commit-Queue: Stéphane Marchesin <marcheu@chromium.org>
Tested-by: Stéphane Marchesin <marcheu@chromium.org>