UPSTREAM: regulatory: allow VHT channels in world roaming
Upstream commit 8ab9d85c650477cbf28a0a0dfc9796c505d03d82.

For VHT, the wider bandwidths (up to 160 MHz) need
to be allowed. Since world roaming only covers the
case of connecting to an AP, it can be opened up
there, we will rely on the AP to know the local
regulations.

BUG=chrome-os-partner:20691
TEST=tested that connectivity works over ath9k and iwl7000

Change-Id: I27311a3ad1e85c20dfd65405e01c2e7cdcd8b05a
Acked-by: Luis R. Rodriguez <mcgrof@do-not-panic.com>
Signed-off-by: Johannes Berg <johannes.berg@intel.com>
Reviewed-on: https://gerrit.chromium.org/gerrit/59778
Reviewed-by: Paul Stewart <pstew@chromium.org>