CHROMIUM: i2c-core: dt: Pick i2c bus number from i2c alias if present
This allows you to get the equivalent functionality of i2c_add_numbered_adapter()
with all data in the device tree and no special case code in your driver.
This is a common device tree technique.

For quick reference, the FDT syntax for using an alias to provide an
ID looks like:
aliases {
  i2c0 = &i2c_0;
  i2c1 = &i2c_1;
};

Change-Id: Ia9d4ac35910f2ec5a5a1cbdfc9dd874069ca6381
Signed-off-by: Doug Anderson <dianders@chromium.org>
CC: Mark Brown <broonie@opensource.wolfsonmicro.com>
Signed-off-by: R. Chandrasekar <rcsekar@samsung.com>