drm/exynos: add mixer apply function
Currently, an enabled flag for each mixer window is being set
in both mixer and the common hdmi framework. This patch adds
an apply function in mixer and moves the enabled flag inside
the mixer. This is required to ensure the enabled flag is updated
by the resume flag during dpms.

Signed-off-by: Prathyush K <prathyush.k@samsung.com>