drm/exynos: fimd: add complete_scanout function
The wait_for_vblank interface is modified to the complete_scanout
function in fimd. This patch adds the fimd_complete_scanout function

Inside fimd_complete_scanout, we ensure that a framebuffer with a given
dma-address can be freed without causing a page fault. This is done in
two steps:
1> if the base register has that dma-address then the dma will start
reading from that address in the next vsync. So disable the window.
2> if the shadow register has that dma-address then we need to wait
for the next vsync so that the dma is finished. When the next vsync
comes, either the dma will start reading from a new address or the
window will be disabled. So we can free the framebuffer.

Change-Id: I1fa7d384c61e52a25bc280c8d3d74986603e338a
Signed-off-by: Prathyush K <prathyush.k@samsung.com>