UPSTREAM: watchdog: ath79_wdt: get register base from platform device's resources
The ath79_wdt driver uses a fixed memory address
currently. Although this is working with each
currently supported SoCs, but this may change
in the future. Additionally, the driver includes
platform specific header files in order to be
able to get the memory base of the watchdog
device.

The patch adds a memory resource to the platform
device, and converts the driver to get the base
address of the watchdog device from that.

BUG=chrome-os-partner:24222
TEST=With series /dev/watchdog works and reports watchdog reboots

Change-Id: Icc86199401f315285089204aa1acd3954e9fbf24
Signed-off-by: Gabor Juhos <juhosg@openwrt.org>
Signed-off-by: Wim Van Sebroeck <wim@iguana.be>
Signed-off-by: Doug Anderson <dianders@chromium.org>
(cherry picked from commit 09f5100a592d11dad06b218f41d560ff1f87f666)
Reviewed-on: https://chromium-review.googlesource.com/177916
Reviewed-by: Sonny Rao <sonnyrao@chromium.org>