CHROMIUM: test_nx: make test_nx work again
This makes test_nx work again (has been broken 2.6.28). Can't upstream at
the moment, since upstream (after 3.6) changed module exception handlers
to be relative, which makes it impossible to represent stack and heap
exceptions on 64-bit. For now, though, we are able to use it for
autotest to evaluate our x86 kernels.

BUG=chromium-os:37746
TEST=link build

Change-Id: Ia1c322c420b2f51ed919df03397d13278895023a
Signed-off-by: Kees Cook <keescook@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/40981
Reviewed-by: Grant Grundler <grundler@chromium.org>
Reviewed-by: Mandeep Singh Baines <msb@chromium.org>