CHROMIUM: drm/exynos: fbdev: Don't recompute fb size in fbdev_update
The size variable in fbdev_update is used to set the amount of mapped RAM
needed for the framebuffer.  Thus, it is not just the number of pixels
times the bpp; it could be significantly more due to line length cache
alignment and other padding.

Thus, get the size from the gem object itself instead of computing it.

Signed-off-by: Daniel Kurtz <djkurtz@chromium.org>

BUG=none
TEST=splash screen displays correctly w/ no HDMI
TEST=splash screen displays correctly w/ HDMI

Change-Id: I7df74770d49931f98ebd62d71178755a83d6db57
Reviewed-on: https://gerrit.chromium.org/gerrit/40445
Reviewed-by: Daniel Kurtz <djkurtz@chromium.org>
Commit-Queue: Daniel Kurtz <djkurtz@chromium.org>
Tested-by: Daniel Kurtz <djkurtz@chromium.org>