BACKPORT: ALSA: hda - Prefer binding the primary CLFE output
When 5.1 or more multiple speakers with found but not enough DACs are
available, it's better to bind such pins to the DACs of the primary
outputs with the same channels rather than binding to the first DAC
(i.e. the front channel).  For the cases with two speaker pins, it's
rather regarded as front + bass combination, thus it's more practical
to still bind to the front, though.

Signed-off-by: Takashi Iwai <tiwai@suse.de>
(cherry picked from commit 7385df6134888553b5ede71cd573ffe0429e2a80)

BUG=chrome-os-partner:19100
TEST=Verified the working of HDMI audio with both stereo and headset by building
     the ChromeOS image along with entire 3.9 HDMI driver patches and four
     additional patches from drm/i915.