UPSTREAM: watchdog: core: dt: add support for the timeout-sec dt property
Add support for watchdog drivers to initialize/set the timeout field
of the watchdog_device structure. The timeout field is initialised
either with the module timeout parameter value (if valid) or with the
timeout-sec dt property (if valid). If both are invalid the initial
value is unchanged.

BUG=chrome-os-partner:24222
TEST=With series /dev/watchdog works and reports watchdog reboots

Change-Id: I6c3156f68afb62857459a103e732194b8b76c017
Signed-off-by: Fabio Porcedda <fabio.porcedda@gmail.com>
Acked-by: Nicolas Ferre <nicolas.ferre@atmel.com>
Signed-off-by: Wim Van Sebroeck <wim@iguana.be>
Signed-off-by: Doug Anderson <dianders@chromium.org>
(cherry picked from commit 3048253ed957fc6cdc34599178408559aa1e0062)
Reviewed-on: https://chromium-review.googlesource.com/177899
Reviewed-by: Sonny Rao <sonnyrao@chromium.org>